refactor(test): import from `/src/index.ts`
